feat: 입력 금액 객체 Money 추가 및 VendingMachine에서 책임 이관
